requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
supabase
google-generativeai
openai
stripe
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
aiofiles
apscheduler